+++ CHANGELOG.rst
@@ -2,6 +2,11 @@ Changelog
 =========
 
 
+- Add ``Zulu.datetimetuple()``.
+- Add ``Zulu.datetuple()``.
+- Remove ``Zulu.__iter__`` method. (**breaking change**)
+
+
 v0.11.0 (2017-06-28)
 --------------------